+++ lib/src/condition_variable.dart
@@ -6,23 +6,23 @@ import 'dart:async';
 
 import 'package:meta/meta.dart';
 
-/// A [Condition] allows micro-tasks to [wait] for other micro-tasks to
+/// A [ConditionVariable] allows micro-tasks to [wait] for other micro-tasks to
 /// [notify].
 ///
-/// [Condition] is a concurrency primitive that allows one micro-task to
+/// [ConditionVariable] is a concurrency primitive that allows one micro-task to
 /// wait for notification from another micro-task. The [Future] return from
 /// [wait] will be completed the next time [notify] is called.
 ///
 /// ```dart
 /// var weather = 'rain';
-/// final condition = Condition();
+/// final cond = ConditionVariable();
 ///
 /// // Create a micro task to fetch the weather
 /// scheduleMicrotask(() async {
 ///   // Infinitely loop that just keeps the weather up-to-date
 ///   while (true) {
 ///     weather = await getWeather();
-///     condition.notify();
+///     cond.notify();
 ///
 ///     // Sleep 5s before updating the weather again
 ///     await Future.delayed(Duration(seconds: 5));
@@ -31,12 +31,12 @@ import 'package:meta/meta.dart';
 ///
 /// // Wait for sunny weather
 /// while (weather != 'sunny') {
-///   await condition.wait;
+///   await cond.wait;
 /// }
 /// ```
 // TODO: Apply `final` when language version for this library is bumped to 3.0
 @sealed
-class Condition {
+class ConditionVariable {
   var _completer = Completer<void>();
 
   /// Complete all futures previously returned by [wait].
@@ -55,7 +55,7 @@ class Condition {
   /// the future will be completed, and any new calls to [wait] will return a
   /// new future. This future will also be unresolved, until [notify] is called.
   ///
-  /// The [Future] return from this condition will never throw.
+  /// The [Future] return from this condition variable will never throw.
   Future<void> get wait {
     if (_completer.isCompleted) {
       _completer = Completer();