+++ lib/src/notifier.dart
@@ -6,23 +6,23 @@ import 'dart:async';
 
 import 'package:meta/meta.dart';
 
-/// A [ConditionVariable] allows micro-tasks to [wait] for other micro-tasks to
+/// A [Notifier] allows micro-tasks to [wait] for other micro-tasks to
 /// [notify].
 ///
-/// [ConditionVariable] is a concurrency primitive that allows one micro-task to
+/// [Notifier] is a concurrency primitive that allows one micro-task to
 /// wait for notification from another micro-task. The [Future] return from
 /// [wait] will be completed the next time [notify] is called.
 ///
 /// ```dart
 /// var weather = 'rain';
-/// final cond = ConditionVariable();
+/// final notifier = Notifier();
 ///
 /// // Create a micro task to fetch the weather
 /// scheduleMicrotask(() async {
 ///   // Infinitely loop that just keeps the weather up-to-date
 ///   while (true) {
 ///     weather = await getWeather();
-///     cond.notify();
+///     notifier.notify();
 ///
 ///     // Sleep 5s before updating the weather again
 ///     await Future.delayed(Duration(seconds: 5));
@@ -31,31 +31,33 @@ import 'package:meta/meta.dart';
 ///
 /// // Wait for sunny weather
 /// while (weather != 'sunny') {
-///   await cond.wait;
+///   await notifier.wait;
 /// }
 /// ```
 // TODO: Apply `final` when language version for this library is bumped to 3.0
 @sealed
-class ConditionVariable {
+class Notifier {
   var _completer = Completer<void>();
 
-  /// Complete all futures previously returned by [wait].
+  /// Notify everybody waiting for notification.
   ///
-  /// Calls to [wait] after this call, will not be resolved, until the next time
-  /// [notify] is called.
+  /// This will complete all futures previously returned by [wait].
+  /// Calls to [wait] after this call, will not be resolved, until the
+  /// next time [notify] is called.
   void notify() {
     if (!_completer.isCompleted) {
       _completer.complete();
     }
   }
 
+  /// Wait for notification.
+  ///
   /// Returns a [Future] that will complete the next time [notify] is called.
   ///
-  /// This will always return an unresolved [Future]. Once [notify] is called
-  /// the future will be completed, and any new calls to [wait] will return a
-  /// new future. This future will also be unresolved, until [notify] is called.
-  ///
-  /// The [Future] return from this condition variable will never throw.
+  /// The [Future] returned will always be unresolved, and it will never throw.
+  /// Once [notify] is called the future will be completed, and any new calls
+  /// to [wait] will return a new future. This new future will also be
+  /// unresolved, until [notify] is called.
   Future<void> get wait {
     if (_completer.isCompleted) {
       _completer = Completer();