+++ test/stream_queue_test.dart
@@ -250,7 +250,8 @@ void main() {
       var skip4 = events.skip(1);
       var index = 0;
       // Check that futures complete in order.
-      Func1Required<int?> sequence(expectedValue, sequenceIndex) => (value) {
+      Func1Required<int?> sequence(int expectedValue, int sequenceIndex) =>
+          (value) {
             expect(value, expectedValue);
             expect(index, sequenceIndex);
             index++;
@@ -390,7 +391,7 @@ void main() {
     });
 
     test('forwards to underlying stream', () async {
-      var cancel = Completer();
+      var cancel = Completer<int>();
       var controller = StreamController<int>(onCancel: () => cancel.future);
       var events = StreamQueue<int>(controller.stream);
       expect(controller.hasListener, isFalse);
@@ -872,7 +873,7 @@ void main() {
 
       // Regression test.
       test('pending child rest requests emit no more events', () async {
-        var controller = StreamController();
+        var controller = StreamController<int>();
         var events = StreamQueue(controller.stream);
         var transaction = events.startTransaction();
         var queue = transaction.newQueue();
@@ -906,7 +907,7 @@ void main() {
       });
 
       test('before the transaction emits any events, does nothing', () async {
-        var controller = StreamController();
+        var controller = StreamController<int>();
         var events = StreamQueue(controller.stream);
 
         // Queue a request before the transaction, but don't let it complete
@@ -1002,7 +1003,7 @@ void main() {
       });
 
       test('before the transaction emits any events, does nothing', () async {
-        var controller = StreamController();
+        var controller = StreamController<int>();
         var events = StreamQueue(controller.stream);
 
         // Queue a request before the transaction, but don't let it complete