+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## 2.12.0-wip
 
 - Require Dart 2.19
-- Add `Condition`, allowing tasks to wait for `Condition.wait`, which is
-  completed the next time `Condition.notify` is called by another task.
+- Add `ConditionVariable`, allowing tasks to wait for `ConditionVariable.wait`,
+  which is completed the next time `ConditionVariable.notify` is called.
 - Added `StreamExtensions.boundedForEach(N, each)` to enable concurrently
   running a micro-task for each item in a stream, while never running more
   than `N` micro-tasks concurrently.